requirements.txt
aiobotocore==2.4.2
aiofiles==23.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aioitertools==0.12.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
botocore==1.27.59
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
fastapi==0.115.12
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
jmespath==1.0.1
multidict==6.2.0
numpy==2.2.4
openai==1.68.2
propcache==0.3.0
pydab==0.0.0.dev0
pydub
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soundfile
starlette==0.46.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==1.26.20
uvicorn==0.27.1
wrapt==1.17.2
yarl==1.18.3
webrtcvad == 2.0.10
google-genai
dotenv
pillow
regex